test: Merge modules for fracture intersection tests
